Moved bunyan stdout configuration into config.json
An attempt to configure stdout and stderr streams in JSON and pass it
directly to bunyan was not successful. A related, still open discussion
can be found at

https://github.com/trentm/node-bunyan/issues/153

This PR suggests a workaround: create a JSON object from the 'raw'
configuration, substituting 'process.stdout' and 'process.stderr'
strings with corresponding objects, and then pass it to bunyan.