Do not rely object prototypes when evaluating 'code' error property
Before this fix the ErrorContext class used the following approach: in
order to check whether _err.code field is meaningful, we checked for
the prototype first (so we could iterate over property list without
exceptions), and did this search then. However this approach is wrong
and gives incorrect results for different object prototypes (which can
be arbitrary specified with Object.setPrototypeOf() and can differ from
Object.prototype). Moreover, we do not even search through property
list, because the only assumed class with 'code' property in this
context is "HttpError".