Supply staged commits with required status checks
Without this, GitHub cannot detect any association between a staged
commit and the corresponding PR, forbidding applying this commit to the
base branch (when "Include administrators" checkbox in the protected
branch settings is on).

With this fix, if a staged commit has all checks passed(staging_checks
parameter), we assign a missing "required" check(s) to it, actually
duplicating a passed check. This passed check is simply selected by
the beginning of its 'context' field. For example, if a passed check for
a commit has "Jenkins d-ubuntu-zesty" context, and the protected base
branch has "Jenkins" required check configured, we create "Jenkins"
passed check for the commit, referencing test results from
"Jenkins d-ubuntu-zesty" context.