The boot should process PRs with 'M-cleared-for-merge' first
Before this change, there could be one inconvenience if the bot is
configured with 'guarded_run' option.  A user marked his PR 'Y' with
'M-cleared-for-merge', but the bot started PR 'X', because X < Y (both
'X' and 'Y' can be ready from the bot point of view). So, the
bot ends up waiting for 'Y' on 'guarded_run' but the user is upset not
seeing bot's activity on the expected PR.