Do not create duplicated statuses for staged commit (#22)
Anubis was copying non-required checks to the staged commit but not
seeing those copied checks in the staged commit, resulting in an
ever-increasing number of checks in the staged commit (until the GitHub
limit of 1000 checks was reached). This bug was caused by 8510a56.

Also: fixed several method name inconsistencies in GitHub wrappers
(debug information).