guarded_run: let cleared-for-merge PRs overrule the current one
This is another attempt, the first one (rejected due to complexity) was
done at 1eeb3ed. In this new approach all the information about
selecting a next PR for for processing (including the current PR) is
stored inside the sorted PR list. The sort order is based on the
following conditions, by priority:

* PR marked with M-cleared-for-merge (for guarded_run only)
* PR which is 'current'
* PR with a lower number