Further refactoring
The major change here is a new 'post-staged' processing state (in
addition to 'pre-staged' and 'staged'), which starts just after
successful fast-forwarding into the base branch.  During this step all
required PR cleanup should be done atomically. The difference between
first two state and this new one is in that how errors/exceptions are
handled.  If there is an exception during pre-staged/staged, the
higher-level code (PrMerger.js) just starts another PR.  For post-staged
exceptions, the PR is considered as 'suspended', and does not start
another PR, until all required cleanup is completed.

The PR state (and related information) is now calculated just after
PR code gets control. This allowed to eliminate confusion and
code duplication (e.g., with _tagIsFresh()).

A special care is taken over possible 'half-cleanupped' PRs, when an
exception occurs during post-staged. In this case, the code should aware
that the requested PR list does not contain this PR and 'mergeable'
status is not calculated for this PR.