Properly handle getReference() API call results (#28)
Before this fix, Anubis assumed that getReference() would return the
result only on exact ref matching. This is is not true: if the ref doesn't
exist in the repository, but there are other refs starting with the
ref, they are returned as an array. This resulted in an unhandled
exception within the promise and getting stuck on the unresolved
promise. We should explicitly reject such wrong results.

Also, I added a process unhandledRejection event handler to make it
easier to detect similar bugs in the future.