Inform run() about PR-specific exceptions
runStep() was not informing the caller about PR-specific exceptions
unless the exception affected the very last PR. That inconsistency did
not make sense to me. As for the fix, I was not sure whether to ignore
all PR-specific exceptions or infom run() that there were some
PR-specific exceptions.

Since a PR-specific exception might be caused by some bad global state,
it feels appropriate to inform run() so that run() can reset as much of
that global state as possible before repeating the attempt. I have
implemented that approach.