Update staged commit approval status once, in update()
We used to call this._setApprovalStatus(this._tagSha) twice, once when
creating a staged commit and once when trying to advance that staged
commit towards merging. Since we now always proceed from one stage to
the next, there is no need to set (presumably the same) statuses twice.

Also, _setApprovalStatus() is a form of a PR maintenance (rather than
advancement) so the call belongs to update(). It only makes sense for
the staged PR, but a this._tagSha check should take care of that.