Polished PR state terminology
Dealing with "pre-conditions for post-staging a PR" was a bit awkward,
and "merged" is what we really mean by "post-staged".

"Pre-stage" made me think that we always move into one direction, from
"pre" to "post". However, staged PRs often go _back_ to pre-staged. The
standard git term for an equivalent area is "working directory". The
files in a similar state are sometimes (unofficially) called "unstaged".
Calling a PR "working" would be confusing and using the reverse of
"staged" is only appropriate in some cases. "Brewing" is not ideal, but
that is the best I could come up with.

Also marked a suspected problem with current().