Hid re-processing() logic from run() step, addressing a branch TODO
Also do not assert if somebody changes PR title while we are
re-processing(). Re-processing is nice, but I think it is better to give
it one shot than to have a scary reprocess while you want loop or a
slightly more complex reprocess a few times loop. We may come back to
this decision later if reprocessing actually happens a lot.

Also simplified StepResult while adding "pr is/was staged" flag to it.
TODO: Succeed() and Suspend() can be merged into one "no delay" method
or simply removed (with no delay by default).