Clarified tag deletion conditions in process()
Before this refactoring it was not easy to comprehend conditions related
to failed tag deletion. The already existing exceptions table has this
information we should refer to.  Also made dryRun() a separate condition
to be in line with other dryRun() usecases (e.g., it would be easier to
refactor this code if we decide to move dryRun into API methods
themselves).