Always treat the staged commit as stale if the base branch changes
Before this change the staged commit with failed checks was still kept
in order to avoid useless rebuilds (see the corresponding TODO). However
this approach could cause problems (maybe rare), because the bot would
igore such PRs in case the error has been fixed on the base branch.  So
it is better to give up this optimization in favor of faultless behavior.

While fixing this I had to refactor our branch-comparing code, which was
not clear and had some duplication. The new BranchComparator class now
is responsible for commits comparing.