Polished BranchComparator documentation
The "there are new commits on base since the common ancestor with the
branch" definition of merged() was wrong because the same can be said
about the diverged case.

I also find the "common ancestor" references a bit confusing (and
unnecessary) because when somebody merges base changes into a feature
branch, it becomes difficult for them to be sure what that "common
ancestor" is as they see "identical" commits in both branches that are
not really a common ancestor. Technically, it is all just a comparison
of commit sets/chains -- which chain fully contains the other chain, but
it is difficult for humans to think that way.

I also did not like bare/unqualified "branch" use because "base" is also
a branch, but I did not rename branchRef.