Renamed BranchComparator to BranchPosition
It was awkward to read expressions like _stagedToBaseComparator.ahead()
because they looked like "comparator is ahead". The class is primarily
used as a position (or something like that), not a comparator. The fact
that we have to explicitly call the compare() method to initialize that
position is just an implementation nuance/annoyance (which this commit
does not address) -- it should have been a function.