Set M-failed-staging-checks if staging failed
Apart from setting this label on failure in staged state, we also need
it if a failure was discovered when calculating PR state.  Before this
fix, failed staging checks forced PR to become brewing, new commit
creation, failing and repeating the cycle from the beginning.

Also removed a stale member.