Properly handle missing target_url in statuses
An identified case, when target_url field is empty, is
the initial Jenkins 'pending' status, when the test was
queued but not started yet. Instead of checking the url
upon receiving the status, now it is checked upon its usage,
i.e., when the required status is copied into the staged
commit statuses.