Added a new M-failed-staging-other label
This new label is set by the bot but cleared only by a human.  Its
purpose is to prevent PRs from 'livelocking' when an unexpected error
occurs during the staging phase. For example, if we have 2 PRs and one
of them experiences such an error, the control passes to the second PR
which (after creating a staged commit), gets into the same error,
passing the control back to the first PR and so on.