Added a new M-failed-staging-other label (#36)
This new label is set by the bot but cleared only by a human.  Its
purpose is to prevent PRs from 'livelocking' if an unexpected error
occurs during the staging phase. For example, if we have PRs A and B
and A experiences such an error, the control passes to B
which (after creating the staged commit), gets into the same error,
passing the control back to A and so on.

Also always reset labels ASAP (i.e., after obtaining them from
GitHub).  This approach better reflects the nature of the
'temporary' labels: on each pass the bot should calculate them from
scratch (and leave them unset if an error occurred before the
calculation). This does not prevent the bot from lying about the
PR state. For example, failedDescriptionLabel may be removed while the
message is still invalid. However, it looks preferable than keeping the
label (confusing the user) if the description has been fixed already.

Also clarified descriptions of 'temporary' labels.