Abort stalled GitHub transactions (#42)
The bot may wait for a GitHub response for too long (several hours),
reacting to notifications but stalling progress on all PRs. We need a
timeout to avoid such delays. Fortunately, the octokit/rest library
provides the 'timeout' parameter (which defaults to no timeout). A
timed-out API call returns a 504 error, and we handle that correctly.

TODO: Eventually, we may add a more general "PR processing" timeout, but
this immediate solution is too simple to pass up.

Addresses issue #41.