Ignore M-failed-staging-checks when staging checks succeed
This label indicates that staging checks failed some time ago. The bot
ignored PRs with this label but only if the staging commit checks were
still broken. On the other hand, if these checks restart and succeed,
the bot assumes that the staging commit is OK and tries to merge
it (not caring of the label), ending up with the merged PR with
wrong label. So we need keep the label and staging statuses in sync.