Removed the GitHub commits search optimization for master
Github Search API has a custom rate limit, which is 30 requests per
minute for our token authorization. This limitation did not matter when
there were few open PRs. Now, as the PR list has grown, the limit was
reached on every pass (according to the bot logs) and several
unfortunate (random) PRs were constantly marked with M-failed-other.

Fortunately, we can do without this advanced Search API and use the
standard GitHub 'List commits' API. There is a couple of disadvantages
with this approach:

1. It does not allow to filter by commit message text on GitHub.
   However, since normally all master messages conform to the search
   pattern, dropping it should basically yield the same results.
2. It is a 15-20% slower. However, taking into account the 30 per-minute
   limitation, the final PR scan time would be almost the same (with the
   difference that Anubis would wait until GitHub resets its counters).

So, since the benefits of the advanced search are small versus
implementation complexity, we decided to drop it for good.