Use commit creation time when calculating staged commit freshness
Practice showed that if a GitHub commit brings no changes, GitHub
does not regenerate the tree object of the PR merge commit but
re-creates merge commit with the same tree object. In this case,
if the PR had the staged commit already, the bot deemed it to be fresh
because it compared commit tree SHAs. This behavior could confuse the
user who assumed that any new PR commit should force it restart from
scratch.