Re-stage PRs after empty PR branch commits (#48)
After an empty PR branch commit, GitHub does not regenerate the tree
object of the PR merge commit. GitHub creates a new merge commit with
the old tree object instead. In this case, if the PR had a staged commit
already, the bot deemed it to be fresh because the staged commit and the
GitHub merge commit had the same commit tree (SHA). Users were naturally
surprised by the lack of a staged commit regeneration after a branch
change, especially since empty branch commits are usually used to
"awake" some automation, like CI stuck on an older (staged) commit.

This change was triggered by the following discussion:
https://github.com/squid-cache/squid/pull/651#issuecomment-639570539