Do not uselessly flip M-abandoned-staging-checks
When this label was added at 5ef56f6, with the intention to make it
short living (i.e., during a single PR processing cycle).
It is set (on GitHub) when the staged PR becomes stale (probably due to
new branch commits) and then immediately unset, assuming
that a fresh PR staged commit is going to be always created
at the same cycle (thus reflecting the 'abandoning' fact in the PR
history).

However, this assumption was wrong, because the former-staged PR is not
always immediately promoted after new branch commits (e.g., a review was
requested).  As a result, the label was constantly switched on
and off after each related PR event. We should drop the label only after
a new staged commit has been actually created.