Do not set abandonedStagingChecksLabel() twice
Commit 5ef56f6 set abandonedStagingChecksLabel() immediately upon
discovering abandonment conditions. Immediate setting was probably
deemed necessary because, otherwise, the plan to set the label would be
erased by a _removeTemporaryLabels() call that happens _before_ we check
staging preconditions. If staging preconditions fail, then a new staged
commit is not created but the abandonedStagingChecksLabel() is already
gone. And if they fail _silently_[1] (i.e. via _exObviousFailure()),
then the user may be confused to see an all-green staged commit that is
not going anywhere -- the fact that the commit is stale/abandoned may
not be clear to the user in that case (which is the original problem
that commit 5ef56f6 was trying to solve).

But such an immediate setting is also a bad idea because the label may
not be necessary: If the existing "clear all labels then set the still
needed ones" logic is bypassed, we get excessive labeling (which is the
problem this branch tries to fix).

The combination of the two restrictions means that
abandonedStagingChecksLabel() must be set after
_removeTemporaryLabels(), but we should not do it in _update() because a
successful _createStaged() would have to overwrite and we cannot do it
in _createStaged() because that method may not be reachable due to
_checkStagingPreconditions() failure.

I see only one solution -- set abandonedStagingChecksLabel() in
process() general cleanup code. That code runs for all unmerged PRs (and
merged PRs do not need the label).

[1] We could get rid of abandonedStagingChecksLabel() entirely and make
_exObviousFailure()s more "obvious" via labeling. However, I am worried
that it would introduce a lot of labeling noise. There are just too many
_exObviousFailure() calls, many of those failures may happen without a
stale staged commit, and some are probably really obvious to some users
(that will complain about or be annoyed with excessive labeling). In a
sense, that is the whole idea behind abandonedStagingChecksLabel(): It
tells the user that Anubis is no longer looking at the stale staged
commit, _regardless_ of everything else that is going on with the PR
state at the same time; it does not tell "why", reducing all the
possibilities to one label and forcing users to analyze the PR state.
Some day, this should be done by adding a PR comment instead (which can
detail the reason the staged commit got abandoned and not replaced).