Always compute _stagingFailed from scratch
without loading its value from the label.

The previous approach (with loading from the label) was based
on the two points:

1. An exception in the code before _stagingFailed recalculation may
cause the unexpected label flipping.

2. An already merged (but still open) PR should keep the label intact
(indicating the real state of the merged commit), regardless of the
current PR state (the statuses could change since that).

However, (1) is probably irrelevant, since in this case there would be
another label indicating an even worse problem; and there is probably no
disadvantage in keeping the label up-to-date in (2), because we could
always inspect the statuses of the merged commit manually.

The new approach should be simpler/better overall because there is no
danger of making wrong decisions based on a (possibly stale) label
state via _stagingFailed.