Prevent "API rate limit exceeded" GitHub error for non-search requests
GitHub limits the request rate for an authenticated user as 5000/hour
max.  This limitation may be hit in some circumstances, especially when
the number of open PRs is relatively large (~100 and more), and there
are many GitHub events occurring during the hour, forcing the bot to go
through all open PRs and send GitHub requests - at least one for each
open PR.

To overcome this limitation, we pause each API call after reaching a
hard-coded requests limit (1000 for now) within the hour. We expect that
in most cases this 1000 limit will not be reached - no delay will be
imposed and the bot will work as before. The delay is calculated based
on GitHub "x-ratelimit*" values, sent back to client in each response.