Simplified, removing InvalidFieldsTokenizer
Actually, this class looks like overengineering since we do not need to
iterate over all invalid tokens but rather find the first occurrence of
it. A single method should be fine to do this.

Also fixed the regex determining the trailer boundary. We need to allow
trailers starting with any attribute, not only Co-authored-by one.