Removed _stagingFailed flag
This flag is problematic in some aspects:

* it duplicates _stagedStatuses.failed() when staged commit exists
* it has too broad scope to be used, e.g., in
 _checkStagingPreconditions() context (where only 'abandoned' commit
   is expected)
* it introduces some difficulties when distinguishing its
  'undefined' and 'false' values

To avoid these problems, I decided store abandoned staged checks
instead. On one hand, the narrow scope of this field allows to use it
in specific contexts. On the other hand, we can inspect both abandoned
and staged checks in a broader scope (e.g., when labels are applied).