fixup: Simplified messageValid debugging
I could not stomach a "paranoid null check" that did not really check
the "undefined or non-null" invariant it was stating. There is no code
that assigns null to this._commitMessage, so I think we can rely on an
undefined check for debugging purposes.

If we do need a paranoid check, we should add a matching assert().