Convert Authored-By statement into commit author (#53)
If the PR description starts with "Authored-By:" attribute, the
specified author credentials are used as the commit author instead of
the ones provided by GitHub (i.e., the PR author). The entire
attribute line (including the trailing empty lines) is removed from
the PR message.

If PR author credentials cannot be parsed, the PR is rejected as having
an invalid description. The PR is rejected also if more than one
'Authored-by' attribute is specified.

Similar checks are applied also to 'Co-authored-by' attributes (coming
in the trailer part of the PR message). The credentials part of
this attribute follows the same syntax, but the trailer may have more
than one 'Co-authored-by'.