Do not rely on dates when determining the abandoned staged commit
Since staged and merge commit are created independently by the bot and
GitHub, it is possible that the bot creates staged commit based on the
(already) stale merge commit, and we cannot for sure determine which
of them is newer simply comparing their dates. However, it is much
simpler to just compare tree SHAs - match indicates that the abandoned
commit exists.