Support GitHub check-runs API (#58)
These GitHub-driven checks are collected and analyzed
the same way as the (old) commit statuses. For simplicity,
check-runs are stored in the same StatusChecks array.

The 'staging_checks' parameter now reflects the total
number of required commit statuses and required GitHub check-runs.

This support is necessary for projects that
require GitHub Actions.