Wait if staging branch status is being updated (#61)
Generally we expect that if a GitHub API call returns success,
the operation has been completed on GitHub. However, this proved
to be wrong at least once for updateReference() call.

It is still an open question whether this problem relates to just
updateReverence() or all modifying operations (with PATCH/POST method).
We would need to come up with a more general approach in the second
case.