Ignore irrelevant check-runs (#62)
Check Runs API returns the whole set of completed checks for the given
commit. In most cases the set does not contain duplicates because these
checks start only once for the commit. However, sometimes the same
commit may be tested multiple times (e.g., when the PR branch is
force-pushed with the same PR head commit). In this case, all results
are returned, making Anubis assert when checking tests uniqueness. To
avoid this, we need to select only the latest check runs from GitHub.