Properly disginguish PR SHAs and PR branches
We cannot rely on the 40-character check because branches can be easily
a string of that length, but in a format allowed by GitHub (i.e., not in
the '40 characters containing only 0-9 and A-F' format).  The
new PrId class should provide the required functionality to
distinguish three PR ID types: SHAs, branches and PR numbers.