Ignore irrelevant check-runs for staged commits
A similar problem for PR branch was fixed in 43670f1.  Unlike PR branch,
staged commit duplicates occur not because of a force-push to a branch
but due to the non-atomic checks retrieval method, with artificial
delays between fetched pages. If a sent check's status changes during
the retrieval, it may be sent the second time in another page with the
updated status. Now we select latest check runs for staged commits,
as we already do for PR branch.