Always treat the currentPr as 'updated'
Otherwise, if there has been no activity on the currentPr, it will be
ignored by the optimization, its staged commit will be abandoned and the
bot proceeds with creating a staged commit for another PR (afterwards
repeating this scenario again and again). This fix covers the
case when the currentPr is not marked with M-cleared-for-merge.