Ignore irrelevant branches in the status event handler
A status event may contain several branches when the commit belongs to
several branches. For example, if a PR X was merged and a new branch
has been created, the head of this branch will also contain the '...#X'
message. The status event hander misattributes then this branch to the
already closed PR X.

The new check is more accurate - we extract the PR number from the
branch if and only if it is the staging branch.