Properly handle adding the same optional status
During _supplyStagingWithPrRequired(), StatusChecks.addOptionalStatus()
asserted when it tried to add a 'copied from PR' status to the
staged StatusChecks array, which already had some status with the same
context. The 'copied from' status should overwrite the existing
staged status. Just before the assertion, Anubis properly creates this
'copied from' status on GitHub - so the fix is about adjusting
StatusChecks to represent the current state of staged checks.

Another problem was that the 'copied from' status was created even if it
already existed.