Applied another solution
The previous (second) solution, based on setting
M-abandoned-staging-checks depending on M-waiting-staging-checks and
M-passed-staging-checks was flawed: the lifetime of the former two
labels is less than that of the latter one. I could not think of any
cheap way of calculating the 'abandoned' state other from requesting PR
events history and searching the referenced commits for the stale staged
commit. Since requesting all events for each processing PR can be
costly, I added an optimization that fetches only several recent
events. In the future we may vary the number of fetched events so that
most PRs were marked correctly.