Applied another solution
This (third) solution is a combination of the previous two: we check the
labels (indicating whether the PR has been ever staged) and only then
acquire PR events (a costly operation). With this optimization we
minimize the number of PRs fetching events and at the same time
guarantee that we are not going to miss some PRs that should be marked
as abandoned.